+++ Changes.md
@@ -62,6 +62,7 @@ Improvements
   - Added `emissionSamplingMethod` plug.
   - Added `isCausticsCaster` and `isCausticsReceiver` plugs.
 - Render : Improved clearing of the compute and hash caches prior to rendering, potentially providing more memory to the renderer. Note that clearing is now only performed in the `execute` and `dispatch` apps.
+- OpenImageIOReader : Added `fileValid` BoolPlug that outputs `True` if a file exists for the current frame.
 
 Fixes
 -----